src/requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.7
ansible==2.8.4
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
clyent==1.2.2
conda==4.7.12
conda-build==3.18.8
conda-package-handling==1.6.0
conda-verify==3.4.2
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.5
Django==2.2.6
ecdsa==0.13.3
entrypoints==0.3
filelock==3.0.12
future==0.17.1
get==2019.4.13
glob2==0.7
google-auth==1.6.3
graphviz==0.13
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
libarchive-c==2.8
lief==0.9.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
oauthlib==3.1.0
olefile==0.46
panda==0.3.1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
ptyprocess==0.6.0
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycosat==0.6.3
pycparser==2.19
pydot==1.4.1
pydotplus==2.0.2
PyFunctional==1.2.0
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-jose==3.0.1
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
QtPy==1.9.0
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sklearn-utils==0.0.15
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.3
sqlparse==0.3.0
statsmodels==0.10.1
tabulate==0.8.5
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.24.2
virtualenv==16.7.6
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
python-twitter
